+++ .eslintrc.js
@@ -8,6 +8,7 @@ module.exports = {
   env: {
     browser: true,
     es2022: true,
+    node: true,
   },
   parser: '@typescript-eslint/parser',
   parserOptions: {
