+++ .vscode/launch.json
@@ -6,14 +6,16 @@
       "command": "pnpm dev-codemirror",
       "name": "Codemirror Playground",
       "request": "launch",
-      "type": "node-terminal",
-      "preLaunchTask": "npm: watch"
+      "type": "node-terminal"
     },
     {
       "type": "extensionHost",
       "request": "launch",
       "name": "VSCode Playground",
       "runtimeExecutable": "${execPath}",
+      "env": {
+        "watch": "true"
+      },
       "args": [
         "--extensionDevelopmentPath=${workspaceRoot}/packages/vscode-extension"
       ],
@@ -25,7 +27,7 @@
       ],
       "autoAttachChildProcesses": true,
       "sourceMaps": true,
-      "preLaunchTask": "npm: build-vscode"
+      "preLaunchTask": "npm: dev-vscode"
     },
     {
       "type": "extensionHost",
