--- packages/vscode-extension/esbuild-extension.js
@@ -1,26 +1,0 @@
-const esbuild = require('esbuild');
-
-const production = process.argv.includes('--production');
-
-async function main() {
-  const ctx = await esbuild.context({
-    entryPoints: ['src/extension.ts'],
-    bundle: true,
-    format: 'cjs',
-    minify: production,
-    sourcemap: !production,
-    sourcesContent: false,
-    platform: 'node',
-    outfile: 'dist/extension.js',
-    external: ['vscode'],
-    logLevel: 'silent',
-    conditions: ['require'],
-  });
-  await ctx.rebuild();
-  await ctx.dispose();
-}
-
-main().catch((e) => {
-  console.error(e);
-  process.exit(1);
-});
