+++ packages/vscode-extension/esbuild-extension.mts
@@ -1,0 +1,23 @@
+import * as esbuild from 'esbuild';
+const production = process.argv.includes('--production');
+
+const ctx = await esbuild.context({
+  entryPoints: ['src/extension.ts'],
+  bundle: true,
+  format: 'cjs',
+  minify: production,
+  sourcemap: !production,
+  sourcesContent: false,
+  platform: 'node',
+  outfile: 'dist/extension.js',
+  external: ['vscode'],
+  logLevel: 'info',
+  conditions: ['require'],
+});
+
+if (production) {
+  await ctx.rebuild();
+  await ctx.dispose();
+} else {
+  await ctx.watch();
+}
